+++ Features/vexo/Dungeons/allLeapedP5.js
@@ -4,7 +4,7 @@ import { onChatPacket } from "../../../../BloomCore/utils/Events";
 import { getAllPlayers } from '../../../utils/util';
 
 let isInP5 = false;
-let notifiedAllLeap = false;
+let notifiedallLeaped = false;
 
 registerWhen(onChatPacket((event) => {
     isInP5 = true;
@@ -14,11 +14,11 @@ registerWhen(register("tick", () => {
     if (getAllPlayers().filter(player => player.y < 20).length === 5) {
         World.playSound("note.pling", 1.0, 2.0);
         Client.showTitle("&fALL LEAPED!", "", 0, 50, 5);
-        notifiedAllLeap = true;
+        notifiedallLeaped = true;
     }
-}), () => config.allLeaped && isInP5 && !notifiedAllLeap)
+}), () => config.allLeaped && isInP5 && !notifiedallLeaped)
 
 register(`worldUnload`, () => {
     isInP5 = false;
-    notifiedAllLeap = false;
+    notifiedallLeaped = false;
 });
