+++ index.js
@@ -1,17 +1,17 @@
 import config from "./config"
 
-import "./Features/vexo/Dungeon-QOL/ffrz"
-import "./Features/vexo/Dungeon-QOL/m7padTimer"
-import "./Features/vexo/Dungeon-QOL/sbTNT"
+import "./Features/vexo/Dungeons/ffrz"
+import "./Features/vexo/Dungeons/m7padTimer"
+import "./Features/vexo/Dungeons/sbTNT"
 import "./Features/vexo/Dev/printEntity"
 import "./Features/vexo/Dev/starredMobEsp"
-import "./Features/vexo/Dungeon-QOL/RagAxeNow"
-import "./Features/vexo/Dungeon-QOL/allLeapedP5"
+import "./Features/vexo/Dungeons/RagAxeNow"
+import "./Features/vexo/Dungeons/allLeapedP5"
 import "./Features/vexo/QOL/rejoin"
-import "./Features/vexo/Dungeon-QOL/StopMageSheepRendering"
-import "./Features/vexo/Dungeon-QOL/healerLeapedpre4"
+import "./Features/vexo/Dungeons/StopMageSheepRendering"
+import "./Features/vexo/Dungeons/healerLeapedpre4"
 
 import { GUICommand } from "./Features/gui"
 import { RewarpCommand } from "./Features/vexo/QOL/rewarp"
-import { TYFRCommand } from "./Features/vexo/Dungeon-QOL/tyfr"
-import { SuckTrapCommand } from "./Features/vexo/Dungeon-QOL/SuckTrap"
+import { TYFRCommand } from "./Features/vexo/Dungeons/tyfr"
+import { SuckTrapCommand } from "./Features/vexo/Dungeons/SuckTrap"
