+++ utils/data.js
@@ -13,6 +13,12 @@ export const data = new PogObject("vexo", {
         scale: 1
     },
 
+    RagAxeAlert: {
+        x: Renderer.screen.getWidth() / 2,
+        y: Renderer.screen.getHeight() / 2 + 10,
+        scale: 1
+    },
+
     padTimer: {
         x: Renderer.screen.getWidth() / 2,
         y: Renderer.screen.getHeight() / 2 + 10,
